refactor(hardware-control): modify tests to reflect changes to hardware controller
