+++ api/tests/opentrons/hardware_control/test_api_helpers.py
@@ -1,11 +1,11 @@
 import pytest
 from opentrons.calibration_storage.types import (
-    DeckCalibration,
     SourceType,
     CalibrationStatus,
 )
 from opentrons.hardware_control.util import DeckTransformState
-from opentrons.hardware_control.robot_calibration import RobotCalibration
+from opentrons.hardware_control.robot_calibration import (
+    RobotCalibration, DeckCalibration)
 
 
 @pytest.mark.ot2_only  # ot3 attitude is always correct
