+++ api/tests/opentrons/hardware_control/test_gripper.py
@@ -2,20 +2,25 @@ from typing import Optional, Callable
 import pytest
 
 from opentrons.types import Point
-from opentrons.hardware_control.robot_calibration import load_gripper_calibration_offset
-from opentrons.hardware_control.instruments import gripper
+from opentrons.hardware_control.instruments.ot3 import gripper
 from opentrons.hardware_control.types import CriticalPoint
-from opentrons.calibration_storage.delete import clear_gripper_calibration_offsets
 from opentrons.config import gripper_config
 from opentrons_shared_data.gripper.dev_types import GripperModel
 
 fake_gripper_conf = gripper_config.load(GripperModel.V1)
 
-FAKE_OFFSET = load_gripper_calibration_offset("fakeid123")
 
 
-def test_config_update():
-    gripr = gripper.Gripper(fake_gripper_conf, FAKE_OFFSET, "fakeid123")
+@pytest.mark.ot3_only
+@pytest.fixture
+def fake_offset():
+    from opentrons.hardware_control.instruments.ot3.instrument_calibration import load_gripper_calibration_offset
+    return load_gripper_calibration_offset("fakeid123")
+
+
+@pytest.mark.ot3_only
+def test_config_update(fake_offset):
+    gripr = gripper.Gripper(fake_gripper_conf, fake_offset, "fakeid123")
     config_to_update = {"z_idle_current": 1.0, "jaw_reference_voltage": 0.5}
     for k, v in config_to_update.items():
         gripr.update_config_item(k, v)
@@ -25,11 +30,13 @@ def test_config_update():
     )
 
 
-def test_id_get_added_to_dict():
-    gripr = gripper.Gripper(fake_gripper_conf, FAKE_OFFSET, "fakeid123")
+@pytest.mark.ot3_only
+def test_id_get_added_to_dict(fake_offset):
+    gripr = gripper.Gripper(fake_gripper_conf, fake_offset, "fakeid123")
     assert gripr.as_dict()["gripper_id"] == "fakeid123"
 
 
+@pytest.mark.ot3_only
 @pytest.mark.parametrize(
     "override,result_accessor",
     [
@@ -48,15 +55,15 @@ def test_id_get_added_to_dict():
 def test_critical_point(
     override: Optional[CriticalPoint],
     result_accessor: Callable[[gripper.Gripper], Point],
+    fake_offset
 ):
-    gripr = gripper.Gripper(fake_gripper_conf, FAKE_OFFSET, "fakeid123")
+    gripr = gripper.Gripper(fake_gripper_conf, fake_offset, "fakeid123")
     assert gripr.critical_point(override) == result_accessor(gripr)
 
-
-def test_load_gripper_cal_offset():
-    clear_gripper_calibration_offsets()
+@pytest.mark.ot3_only
+def test_load_gripper_cal_offset(fake_offset):
     gripr = gripper.Gripper(
-        fake_gripper_conf, load_gripper_calibration_offset("fakeid123"), "fakeid123"
+        fake_gripper_conf, fake_offset, "fakeid123"
     )
     # if offset data do not exist, loaded values should match DEFAULT
     assert (
